Python Code for Gradient Descent
This is a complete code of gradient descent algorithm implemented in Python. This code is for predicting the car price based upon 9 features provided as input.

There are two main files of code,

1. Training:

This file contains the complete code of gradient descent for training the model based upon the data set provided as excel file to the code. For this code, the columns of excel file are ['Car', 'Manufacturing Year', 'MeterReading', 'Transmission', 'Registered City','Color', 'Assembly', 'Engine Capacity', 'Body Type', 'Price']. The last column is of the label which is used later in the algorithm. The final output of this code is the thetas or parameters which are obtained after the training of the model.

2. Prediction:

This file contains simple code of making prediction based upon the hypothesis provided. The detailed hypothesis can be seen inside the code file. Parameters in this file are hard coded and can be changed after training the model on some other data set.

Feel free to use and improve :)